+++ CHANGELOG.md
@@ -11,7 +11,10 @@ Add JSDoc comments to most (hopefully all) code and added changelog
 ## 0.0.3 - NG-Packagr
 
 ### Features
-
 * Move src into src/modules to allow for a demo app
 * Use ng-packagr to help with following best practices
 * Change license from Apache 2.0 to MIT
+
+### BREAKING CHANGES
+* Rename 'getNestedValue' to 'getValue' in 'core/functions/object'
+* Rename 'setNestedValue' to 'setValue' in 'core/functions/object'